% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcode.r
\name{FineTune}
\alias{FineTune}
\title{The ad-AF method involves fine-tuning parameters σ and m based on chosen parameter distributions.}
\usage{
FineTune(
  Sd = c(0.5, 1, 2),
  m = c(370, 740, 1111),
  n,
  pn,
  dist = c("normal", "chisq", "t", "beta", "lnrom"),
  alpha = 0.0027,
  B = 10000
)
}
\arguments{
\item{Sd}{Standard derivation for the CDF of normal distribution: search domains \eqn{\sigma \in \{0.5, 1, 2\}}}

\item{m}{The inflection points for the CDF of normal distribution: search domains \eqn{ m \in \{370, 740, 1111\}}}

\item{n}{Sample size}

\item{pn}{Nominal coverage probability}

\item{dist}{The trial-run distributions are: Normal(0,1), Chi-square(4), t(4), Beta(9,1), and lognormal(0,1) represented by c("normal", "chisq", "t", "beta", "lnorm"), respectively.}

\item{alpha}{False alarm rate}

\item{B}{Number of simulation replications}
}
\value{
A list of the optimal tuning parameters
\itemize{
 \item{ Opt.m:}{ The optimal inflection point}
 \item{Opt.sd:}{ The optimal standard derivation}
 }
}
\description{
The ad-AF method involves fine-tuning parameters σ and m based on chosen parameter distributions.
}
\examples{
FineTune(n = 370, pn = 0.6, dist="t")
}
