% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcode.r
\name{FineTuneWithData}
\alias{FineTuneWithData}
\title{The ad-AF method: fine-tune parameters σ and m based on the given training and testing data.}
\usage{
FineTuneWithData(
  Sd = c(0.5, 1, 2),
  m = c(370, 740, 1111),
  pn,
  Train,
  Test,
  alpha = 0.0027,
  B = 10000
)
}
\arguments{
\item{Sd}{Standard derivation for the CDF of normal distribution: search domains \eqn{\sigma \in \{0.5, 1, 2\}}}

\item{m}{The inflection points for the CDF of normal distribution: search domains \eqn{ m \in \{370, 740, 1111\}}}

\item{pn}{Nominal coverage probability}

\item{Train}{The provided training data}

\item{Test}{The provided testing data}

\item{alpha}{False alarm rate}

\item{B}{Number of simulation replications}
}
\value{
A list of the optimal tuning parameters
\itemize{
 \item{ Opt.m:}{ The optimal inflection point}
 \item{Opt.sd:}{ The optimal standard derivation}
 }
}
\description{
The ad-AF method: fine-tune parameters σ and m based on the given training and testing data.
}
\examples{
Train = matrix(rnorm(10000*370), nrow = 10000 )
Test = matrix(rnorm(10000*370), nrow = 10000 )
FineTuneWithData(Train = Train, Test=Test, pn = 0.7)
}
