% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcode.r
\name{prepro}
\alias{prepro}
\title{Data Preprocessing}
\usage{
prepro(x, trim = TRUE)
}
\arguments{
\item{x}{Input data vector}

\item{trim}{logical; if TRUE, the fraction (0 to 0.01) of observations to be trimmed}
}
\value{
The preprocessed data
}
\description{
Data Preprocessing
}
